+++ server/src/main/java/org/opensearch/action/admin/cluster/health/ClusterHealthRequest.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.action.admin.cluster.health;
 
+import org.opensearch.Version;
 import org.opensearch.action.ActionRequestValidationException;
 import org.opensearch.action.IndicesRequest;
 import org.opensearch.action.support.ActiveShardCount;
@@ -47,6 +48,8 @@ import java.io.IOException;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
+import static org.opensearch.action.ValidateActions.addValidationError;
+
 /**
  * Transport request for requesting cluster health
  *
@@ -55,6 +58,7 @@ import java.util.concurrent.TimeUnit;
 public class ClusterHealthRequest extends ClusterManagerNodeReadRequest<ClusterHealthRequest> implements IndicesRequest.Replaceable {
 
     private String[] indices;
+    private String awarenessAttribute;
     private IndicesOptions indicesOptions = IndicesOptions.lenientExpandHidden();
     private TimeValue timeout = new TimeValue(30, TimeUnit.SECONDS);
     private ClusterHealthStatus waitForStatus;
@@ -90,6 +94,10 @@ public class ClusterHealthRequest extends ClusterManagerNodeReadRequest<ClusterH
         }
         waitForNoInitializingShards = in.readBoolean();
         indicesOptions = IndicesOptions.readIndicesOptions(in);
+        if (in.getVersion().onOrAfter(Version.V_2_5_0)) {
+            awarenessAttribute = in.readOptionalString();
+            level = in.readEnum(Level.class);
+        }
     }
 
     @Override
@@ -118,6 +126,10 @@ public class ClusterHealthRequest extends ClusterManagerNodeReadRequest<ClusterH
         }
         out.writeBoolean(waitForNoInitializingShards);
         indicesOptions.writeIndicesOptions(out);
+        if (out.getVersion().onOrAfter(Version.V_2_5_0)) {
+            out.writeOptionalString(awarenessAttribute);
+            out.writeEnum(level);
+        }
     }
 
     @Override
@@ -269,6 +281,22 @@ public class ClusterHealthRequest extends ClusterManagerNodeReadRequest<ClusterH
         this.level = Objects.requireNonNull(level, "level must not be null");
     }
 
+    public void setLevel(String level) {
+        switch (level) {
+            case "indices":
+                level(ClusterHealthRequest.Level.INDICES);
+                break;
+            case "shards":
+                level(ClusterHealthRequest.Level.SHARDS);
+                break;
+            case "awareness_attributes":
+                level(ClusterHealthRequest.Level.AWARENESS_ATTRIBUTES);
+                break;
+            default:
+                level(ClusterHealthRequest.Level.CLUSTER);
+        }
+    }
+
     /**
      * Get the level of detail for the health information to be returned.
      * Only used by the high-level REST Client.
@@ -277,8 +305,22 @@ public class ClusterHealthRequest extends ClusterManagerNodeReadRequest<ClusterH
         return level;
     }
 
+    public ClusterHealthRequest setAwarenessAttribute(String awarenessAttribute) {
+        this.awarenessAttribute = awarenessAttribute;
+        return this;
+    }
+
+    public String getAwarenessAttribute() {
+        return awarenessAttribute;
+    }
+
     @Override
     public ActionRequestValidationException validate() {
+        if (level.equals(Level.AWARENESS_ATTRIBUTES) && indices.length > 0) {
+            return addValidationError("awareness_attribute is not a supported parameter with index health", null);
+        } else if (!level.equals(Level.AWARENESS_ATTRIBUTES) && awarenessAttribute != null) {
+            return addValidationError("level=awareness_attributes is required with awareness_attribute parameter", null);
+        }
         return null;
     }
 
@@ -290,6 +332,7 @@ public class ClusterHealthRequest extends ClusterManagerNodeReadRequest<ClusterH
     public enum Level {
         CLUSTER,
         INDICES,
-        SHARDS
+        SHARDS,
+        AWARENESS_ATTRIBUTES
     }
 }